Set up forwarders to receive their routing table
The controller will send each node the routing table they should use.
The network topology has been hardcoded into the source code for the
controller so that it has a global view of the network.

In order to enable hostnames to be used in the view of the network, each
forwarder must register with the controller so that its IP address can
be associated with the name given to it, without having to do a reverse
DNS lookup. (That type of thing got messy the last time I tried it).